+++ CHANGELOG.md
@@ -15,9 +15,12 @@ Which is based on [Keep A Changelog](http://keepachangelog.com/)
 
 ### Added
 - client_helpers.rb: Add a simple DockerApi class. Add parse_json method.
+- metrics-docker-container.rb: Friendly names option added
 
 ### Changed
 - metrics-docker-stats.rb: Make use of DockerApi class. Default docker_host defined by DockerApi class. Remove docker_api method.
+- metrics-docker-info.rb: Make use of DockerApi class. Default docker_host defined by DockerApi class. Remove docker_api method.
+- metrics-docker-container.rb: Make use of DockerApi class. Re-enable rubocop for container_metrics method.
 
 ### Fixed
 - metrics-docker-stats.rb: Remove trailing / in name value.